% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationList.R
\name{AnnotationList}
\alias{AnnotationList}
\title{AnnotationList Class.}
\description{
Generate and process an Java ArrayList with Annotation objects.
}
\examples{
if (getOption("bignlp.corenlp_dir") == "") corenlp_install(lang = NULL)

docs <- c("Das ist das erste Dokument.", "Ein weiteres Dokument!")
annoli <- AnnotationList$new(docs)
Pipe <- AnnotationPipeline$new()
Pipe$annotate(annoli)
annoli$as.data.table()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{obj}}{rJava object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnnotationList-new}{\code{AnnotationList$new()}}
\item \href{#method-AnnotationList-as.data.table}{\code{AnnotationList$as.data.table()}}
\item \href{#method-AnnotationList-clone}{\code{AnnotationList$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationList-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationList-new}{}}}
\subsection{Method \code{new()}}{
Initialize AnnotationPipeline
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationList$new(x, purge = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A \code{character} vector with text documents.}

\item{\code{purge}}{A \code{logical} value, whether to remove invalid characters as defined by
\code{corenlp_preprocessing_replacements} before running annotators.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationList-as.data.table"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationList-as.data.table}{}}}
\subsection{Method \code{as.data.table()}}{
Parse annotation output and return \code{data.table}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationList$as.data.table()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnnotationList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AnnotationList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
