% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationPipeline.R
\name{AnnotationPipeline}
\alias{AnnotationPipeline}
\title{AnnotationPipeline Class.}
\description{
Worker behind the higher-level \code{StanfordCoreNLP} class that allows fine-tuned
configuration of an annotation pipeline, see the \href{https://stanfordnlp.github.io/CoreNLP/pipelines.html}{documentation of CoreNLP Pipelines}. The
\verb{$annotate()} method supports processing annotations in parallel. Unlike the
\code{StanfordCoreNLP$process_files()} method for processing the content of files
in parallel, it is a very efficient in-memory operation  and the fastest
option for processsing medium-sized corpora. But as annotations consume a lot
of memory, there are limitations to allocating sufficient heap space required
for the parallel in-memory processing of larger corpora. If heap space is
insufficient, the process may run endless without a telling warning message
or an error. So use the \verb{$annotate()} method with appropriate care.
}
\examples{
A <- AnnotationPipeline$new()
a <- list(
  "This is a sentences.",
  "Yet another sentence."
)
A$annotate(a)
result <- A$as.matrix()

reuters_txt <- readLines(system.file(package = "bignlp", "extdata", "txt", "reuters.txt"))
B <- AnnotationPipeline$new()
B$annotate(reuters_txt)
result <- B$as.matrix()

\dontrun{
library(polmineR)
gparl_by_date <- corpus("GERMAPARL") \%>\%
  subset(year \%in\% 1998) \%>\%
  split(s_attribute = "date") \%>\% 
  get_token_stream(p_attribute = "word", collapse = " ")
C <- AnnotationPipeline$new()
C$annotate(gparl_by_date, 4L)
result <- C$as.matrix()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pipeline}}{AnnotationPipeline}

\item{\code{annotations}}{A \code{List} of \code{Annotation} (Java) objects.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AnnotationPipeline$new()}}
\item \href{#method-annotate}{\code{AnnotationPipeline$annotate()}}
\item \href{#method-as.matrix}{\code{AnnotationPipeline$as.matrix()}}
\item \href{#method-clone}{\code{AnnotationPipeline$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize AnnotationPipeline
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPipeline$new(corenlp_dir = getOption("bignlp.corenlp_dir"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{corenlp_dir}}{Directory where StanfordCoreNLP resides.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-annotate"></a>}}
\if{latex}{\out{\hypertarget{method-annotate}{}}}
\subsection{Method \code{annotate()}}{
Annotate a list of strings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPipeline$annotate(x, threads = NULL, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A list of \code{character} vectors to annotate.}

\item{\code{threads}}{If \code{NULL}, all available threads are used, otherwise an
\code{integer} value with number of threads to use.}

\item{\code{verbose}}{A \code{logical} value, whether to show progress messages.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{List} of \code{Annotation} objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as.matrix"></a>}}
\if{latex}{\out{\hypertarget{method-as.matrix}{}}}
\subsection{Method \code{as.matrix()}}{
Experimental method that will parse annotation output and
return a \code{matrix}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPipeline$as.matrix()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPipeline$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
