% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotatorCoreNLP.R
\name{AnnotatorCoreNLP}
\alias{AnnotatorCoreNLP}
\title{Stanford CoreNLP Annotator Class.}
\description{
Stanford CoreNLP Annotator Class.

Stanford CoreNLP Annotator Class.
}
\examples{
Sys.setenv("_JAVA_OPTIONS" = "")
options(java.parameters = "-Xmx4g")
if (getOption("bignlp.corenlp_dir") == "") corenlp_install(lang = "de")

txt <- "Das ist ein Satz. Und das ist ein zweiter Satz."

CNLP <- AnnotatorCoreNLP$new(
  method = "json",
  corenlp_dir = getOption("bignlp.corenlp_dir"),
  properties_file = corenlp_get_properties_file(lang = "de")
  )
CNLP$annotate(txt = txt)
CNLP$annotate(txt = txt, id = 15L)

CNLP <- AnnotatorCoreNLP$new(
  method = "xml",
  properties_file = corenlp_get_properties_file(lang = "de")
)
CNLP$annotate(txt = txt)

\dontrun{
# Java parallellization
options(java.parameters = "-Xmx4g")
library(polmineR)
library(bignlp)
properties <- list(
  "threads" = "6",
  "annotators" = "tokenize, ssplit, pos, lemma, ner",
  "tokenize.language" = "de",
  "tokenize.postProcessor" = "edu.stanford.nlp.international.german.process.GermanTokenizerPostProcessor",
  "pos.model" = "edu/stanford/nlp/models/pos-tagger/german-ud.tagger",
  "ner.model" = "edu/stanford/nlp/models/ner/german.distsim.crf.ser.gz",
  "ner.applyNumericClassifiers" = "false",
  "ner.applyFineGrained" = "false",
  "ner.useSUTime" = "false",
  "ner.nthreads" = "6"
)
CNLP <- AnnotatorCoreNLP$new(method = "json", properties_file = properties)
merkel <- corpus("GERMAPARL") \%>\%
  subset(speaker == "Angela Merkel" & interjection == "FALSE") \%>\%
  get_token_stream(beautify = TRUE, collapse = " ")
system.time(foo <- CNLP$annotate(merkel))
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tagger}}{Class from Stanford CoreNLP to annotate text}

\item{\code{xmlifier}}{Class from Stanford CoreNLP to generate XML output}

\item{\code{jsonifier}}{Class from Stanford CoreNLP to generate JSON output}

\item{\code{writer}}{Class from Stanford CoreNLP to generate TXT output}

\item{\code{append}}{logical, whether to append output to destfile}

\item{\code{method}}{whith output format to use}

\item{\code{cols_to_keep}}{columns from output to keep}

\item{\code{destfile}}{filename}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AnnotatorCoreNLP$new()}}
\item \href{#method-annotation_to_xml}{\code{AnnotatorCoreNLP$annotation_to_xml()}}
\item \href{#method-annotation_to_json}{\code{AnnotatorCoreNLP$annotation_to_json()}}
\item \href{#method-annotation_to_txt}{\code{AnnotatorCoreNLP$annotation_to_txt()}}
\item \href{#method-annotate}{\code{AnnotatorCoreNLP$annotate()}}
\item \href{#method-clone}{\code{AnnotatorCoreNLP$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatorCoreNLP$new(
  corenlp_dir = getOption("bignlp.corenlp_dir"),
  properties_file,
  method = c("txt", "json", "xml"),
  cols_to_keep = c("sentence", "id", "token", "pos", "ner"),
  destfile = NULL,
  logfile = NULL,
  target = NULL,
  report_interval = 1L,
  gc_interval = 100L
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-annotation_to_xml"></a>}}
\if{latex}{\out{\hypertarget{method-annotation_to_xml}{}}}
\subsection{Method \code{annotation_to_xml()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatorCoreNLP$annotation_to_xml(anno)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-annotation_to_json"></a>}}
\if{latex}{\out{\hypertarget{method-annotation_to_json}{}}}
\subsection{Method \code{annotation_to_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatorCoreNLP$annotation_to_json(anno, id = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-annotation_to_txt"></a>}}
\if{latex}{\out{\hypertarget{method-annotation_to_txt}{}}}
\subsection{Method \code{annotation_to_txt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatorCoreNLP$annotation_to_txt(anno)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-annotate"></a>}}
\if{latex}{\out{\hypertarget{method-annotate}{}}}
\subsection{Method \code{annotate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatorCoreNLP$annotate(txt, id = NULL, current = -1L, purge = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotatorCoreNLP$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
