% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{as.Annotation}
\alias{as.Annotation}
\title{CoreNLP Annotation class.}
\usage{
as.Annotation(x)
}
\arguments{
\item{x}{A \code{data.frame}.}
}
\description{
Generate Annotation class from tokenized data (tabular format).
}
\examples{
\dontrun{
library(polmineR)

p_attrs <- c("word", "pos")
x <- data.frame(lapply(
  setNames(p_attrs, p_attrs),
  function(p_attr) corpus("GERMAPARLMINI") \%>\%
    subset(speaker = "Angela Dorothea Merkel") \%>\%
    get_token_stream(p_attribute = p_attr)
))

a <- as.Annotation(x)

json_outputter <- .jnew("edu.stanford.nlp.pipeline.JSONOutputter")
cat(.jcall(json_outputter, "Ljava/lang/String;", "print", a))

pl <- .jnew("edu.stanford.nlp.pipeline.AnnotationPipeline")
pl$addAnnotator(.jnew("edu.stanford.nlp.pipeline.WordsToSentencesAnnotator", FALSE))
pl$annotate(a)

cat(.jcall(json_outputter, "Ljava/lang/String;", "print", a))
}
}
