% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bignlp-package.R
\docType{package}
\name{bignlp-package}
\alias{bignlp-package}
\alias{bignlp}
\title{R-Package 'bignlp'.}
\description{
R corpus annotation pipeline for large corpora.
}
\details{
The package is an interface to the Stanford CoreNLP Natural Language
Processing Toolkit. CoreNLP is implemented in Java and its "jars" needs to be
available. When the package is attached, the availability of Stanford CoreNLP
is checked. The first take is to use the environment variable CORENLP_DIR. If
it is set, the option bignlp.corenlp_dir is set accordingly. Then we move to
searching for CoreNLP in packages. If it has not been
found, the option bignlp.corenlp_dir is empty ("").
}
\examples{
Sys.setenv("_JAVA_OPTIONS" = "")
if (getOption("bignlp.corenlp_dir") == "") corenlp_install(lang = NULL)
}
\author{
Andreas Blaette
}
\keyword{package}
