% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{corenlp_get_jar_dir}
\alias{corenlp_get_jar_dir}
\title{Get directory with CoreNLP jar files.}
\usage{
corenlp_get_jar_dir()
}
\description{
This utility function checks three options where CoreNLP jars might be: (a) The environment
variable 'CORENLP_DIR', then in the directory 'extdata/corenlp' of the bignlp, and finally,
in the package cleanNLP.
}
