% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenlp.R
\name{corenlp_install}
\alias{corenlp_install}
\title{Install Stanford CoreNLP.}
\usage{
corenlp_install(lang = "de", loc)
}
\arguments{
\item{lang}{Languages to install.}

\item{loc}{Directory where to put jar files. If missing, the files will be
placed in the bignlp package.}
}
\description{
The function provides an installation mechanism to download and
install Stanford CoreNLP within the bignlp package or externally.
}
