% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{corenlp_latest}
\alias{corenlp_latest}
\title{Get version of latest CoreNLP release}
\usage{
corenlp_latest()
}
\description{
Evaluate the website of CoreNLP to get information on the latest CoreNLP
release.
}
