% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conll.R
\name{corenlp_parse_conll}
\alias{corenlp_parse_conll}
\title{Parse CoreNLP CoNLL string output.}
\usage{
corenlp_parse_conll(x, progress = TRUE)
}
\arguments{
\item{x}{A filename, or filenames.}

\item{progress}{logical}
}
\description{
Parse the json output from CoreNLP, either text files, or
NDJSON written to one or multiple files.
}
