% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jvm_is_initialized}
\alias{jvm_is_initialized}
\title{Check whether JVM is initialized.}
\usage{
jvm_is_initialized()
}
\value{
A logical value, \code{TRUE} if JVM has been initialized, \code{FALSE} if
not.
}
\description{
The \code{jvm_is_initialized} utility function checks whether a JVM has been 
initialized already. Parallelisation is likely to fail if this is the case. The 
function issues a warning, if a JVM is already running.
}
