% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mince.R
\name{mince}
\alias{mince}
\title{Mince table into directories with text chunks.}
\usage{
mince(x, dir, chunksize = 10L, progress = interactive())
}
\arguments{
\item{x}{A \code{data.table} with columns 'id' and 'text'. Further columns are
ignored.}

\item{dir}{Superdirectory for directories with text segements that will be
processed sequentially.}

\item{chunksize}{An \code{integer} value, the number of strings that will reside
in the chunk directories.}

\item{progress}{A \code{logical} value, whether to show progress bar.}
}
\value{
The function returns a \code{character} vector with the directories that
contain files with text segments.
}
\description{
Multithreaded processing using the StanfordCoreNLP class requires splitting
up input data into pieces of text ("chunks") available as files that are
processed in parallel. The \code{mince()} function performs this split operation,
i.e. it creates directories with chunks within a superdirectory.
}
\examples{
subcorpora <- corpus("GERMAPARLMINI") \%>\%
  subset(interjection == "speech") \%>\%
  split(s_attribute = "speaker")

speakers <- get_token_stream(subcorpora, collapse = " ", beautify = TRUE, verbose = FALSE)
dt <- data.table(id = 1L:length(speakers), text = unlist(speakers))

mincedtextdir <- tempdir()
dirs <- mince(x = dt, dir = mincedtextdir, chunksize = 10L)
}
