% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{corenlp_parse_json}
\alias{corenlp_parse_json}
\alias{corenlp_parse_conll}
\alias{corenlp_parse_xml}
\title{Parse CoreNLP output formats.}
\usage{
corenlp_parse_json(
  x,
  cols_to_keep = c("sentence", "index", "word", "pos", "ner"),
  output = NULL,
  logfile = NULL,
  progress = TRUE
)

corenlp_parse_conll(x, progress = TRUE)

corenlp_parse_xml(x)
}
\arguments{
\item{x}{A filename, or filenames.}

\item{cols_to_keep}{columns to keep}

\item{output}{a destfile}

\item{logfile}{a character string naming the file to an error log to; if
provided, json strings will be written to this file if parsing the json
string string fails}

\item{progress}{logical}
}
\description{
Parse output from CoreNLP.
}
\details{
The JSON results of applying the Stanford CoreNLP annotators can be
written to a streaming JSON file (ndjson format). \code{corenlp_parse_json}
will parse a json string to a data.frame. If output is specified, the
output will be appended to the file provided. If \code{output} is \code{NULL}, a
data.frame is returned. Strings that cannot be parsed are written to the
logfile, if it is defined. If \code{filename} is present, the function will
process one or more files with the output of Stanford CoreNLP in a NDJSON
format. If the argument \code{output} has been defined during initialization, results are
written/appended to that file. Otherwise, a \code{data.frame} is returned.

\code{corenlp_parse_conll} is based on \code{data.table::fread()}
and supplies settings that prevent undesired behaviour.
}
