% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print.R
\name{corenlp_parse_pretty_print}
\alias{corenlp_parse_pretty_print}
\title{Parse CoreNLP pretty print output.}
\usage{
corenlp_parse_pretty_print(x, mc = 1)
}
\arguments{
\item{x}{a character vector or filename}

\item{mc}{number of cores.}
}
\description{
Parse CoreNLP pretty print output.
}
