% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bignlp-package.R
\docType{data}
\name{reuters_dt}
\alias{reuters_dt}
\title{REUTERS-21578 dataset}
\format{
A \code{data.table} with two 2 and 20 rows:
\itemize{
\item \emph{doc_id}: These are unique \code{integer} values to distinguish documents.
\item \emph{text}: The unprocessed plain text of the documents in the corpus.
}
}
\usage{
reuters_dt
}
\description{
The package includes different representations of an excerpt from the
REUTERS-21578 dataset as sample data. The REUTERS corpus is widely used as
sample data for text classification tasks (Silva, Ribeiro 2010). The data
here is taken from the \emph{tm} package. See files in the 'data-raw' folder of
the package how the sample data has been prepared.
}
\references{
Catarina Silva, Bernardete Ribeiro (2010) \emph{Inductive Inference for Large
Scale Text Classification. Kernel Approaches and Techniques}, Springer:
Berlin, pp. 129ff.
}
\keyword{datasets}
