% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationPipeline.R
\name{AnnotationPipeline}
\alias{AnnotationPipeline}
\title{AnnotationPipeline Class.}
\description{
AnnotationPipeline Class.

AnnotationPipeline Class.
}
\examples{
A <- AnnotationPipeline$new()
a <- list(
  "This is a sentences.",
  "Yet another sentence."
)
A$annotate(a)
anno <- A$annotations$get(0L)
conll_outputter <- rJava::.jnew("edu.stanford.nlp.pipeline.CoNLLOutputter")
y <- do.call(rbind, lapply(
  0L:(A$annotations$size() - 1L),
  function(i){
    read.table(text = conll_outputter$print(A$annotations$get(i)))
  }
))
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{pipeline}}{AnnotationPipeline}

\item{\code{annotations}}{A \code{List} of \code{Annotation} (Java) objects.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AnnotationPipeline$new()}}
\item \href{#method-annotate}{\code{AnnotationPipeline$annotate()}}
\item \href{#method-clone}{\code{AnnotationPipeline$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize AnnotationPipeline
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPipeline$new(corenlp_dir = getOption("bignlp.corenlp_dir"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{corenlp_dir}}{Directory where StanfordCoreNLP resides.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-annotate"></a>}}
\if{latex}{\out{\hypertarget{method-annotate}{}}}
\subsection{Method \code{annotate()}}{
Annotate a list of strings.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPipeline$annotate(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A list of \code{character} vectors to annotate.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{List} of \code{Annotation} objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnnotationPipeline$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
