% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotatorCoreNLP.R
\docType{data}
\name{AnnotatorCoreNLP}
\alias{AnnotatorCoreNLP}
\title{Stanford CoreNLP Annotator Class.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AnnotatorCoreNLP
}
\description{
Stanford CoreNLP Annotator Class.
}
\section{Fields}{

\describe{
\item{\code{tagger}}{Class from Stanford CoreNLP to annotate text}

\item{\code{xmlifier}}{Class from Stanford CoreNLP to generate XML output}

\item{\code{jsonifier}}{Class from Stanford CoreNLP to generate JSON output}

\item{\code{writer}}{Class from Stanford CoreNLP to generate TXT output}

\item{\code{append}}{logical, whether to append output to destfile}

\item{\code{method}}{whith output format to use}

\item{\code{cols_to_keep}}{columns from output to keep}

\item{\code{destfile}}{filename}
}}

\examples{
options(java.parameters = "-Xmx4g")
corenlp_dir <- "/opt/stanford-corenlp/stanford-corenlp-full-2017-06-09"
prop_file <- "/opt/stanford-corenlp/StanfordCoreNLP-german.properties"

txt <- "Das ist ein Satz. Und das ist ein zweiter Satz."

CNLP <- AnnotatorCoreNLP$new(method = "json", corenlp_dir = corenlp_dir, properties_file = prop_file)
CNLP$annotate(txt = txt)
CNLP$annotate(txt = txt, id = 15L)

CNLP <- AnnotatorCoreNLP$new(method = "xml", corenlp_dir = corenlp_dir, properties_file = prop_file)
CNLP$annotate(txt = txt)
}
\keyword{datasets}
