% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_table.R
\name{chunk_table_split}
\alias{chunk_table_split}
\title{Split up file into subsets.}
\usage{
chunk_table_split(input, output = NULL, n, verbose = TRUE)
}
\arguments{
\item{input}{Filename (full path) of a chunk table.}

\item{output}{The output file(s). If \code{NULL}, files will be written to the
session temporary directory.}

\item{n}{The number of (rougly equally sized) files to generate.}

\item{verbose}{Logical, whether to print messages.}
}
\description{
Split up a chunk table file into equally sized files for parallel processing.
}
\details{
The file specified by \code{input} needs to be a single file that exists. It
will be splitted into \code{n} number of subsets. If \code{output} is
\code{NULL}, the files will be generated in the session temporary directory.

The return value is the files that have been written, so that the
\code{chunk_table_split}-function can be used in a pipe.

Note that the input file is assumed to have a header (first line with
colnames). This header will be prefixed to all output files.
}
\examples{
infile <- system.file(package = "bignlp", "extdata", "tsv", "unga.tsv")
outfiles <- chunk_table_split(infile, n = 3L)

dts <- lapply(outfiles, data.table::fread)
sapply(dts, nrow)
}
