% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenlp.R
\name{corenlp_annotate}
\alias{corenlp_annotate}
\title{Annotate a string.}
\usage{
corenlp_annotate(x, destfile = NULL, properties_file, corenlp_dir,
  method = "json", threads = 1L, progress = TRUE, preclean = TRUE,
  verbose = TRUE)
}
\arguments{
\item{x}{input file(s)}

\item{destfile}{output file, or directory, if multiple files are processed in parallel mode}

\item{properties_file}{a properties file to configure annotator}

\item{corenlp_dir}{directory where corenlp resides}

\item{threads}{integer}

\item{progress}{logical}

\item{preclean}{whether to preprocess string}

\item{verbose}{logical}
}
\description{
Use CoreNLP to annotate strings.
}
\examples{
library(data.table)
reuters_dt <- data.table(
  text = readLines(system.file(package = "bignlp", "extdata", "txt", "reuters.txt"))
)
reuters_dt[["id"]] <- 1L:nrow(reuters_dt)

destfile <- tempfile()
corenlp_dir <- system.file(package = "cleanNLP", "extdata", "stanford-corenlp-full-2016-10-31")
properties_english_fast <- system.file(
  package = "cleanNLP", "extdata", "StanfordCoreNLP-english-fast.properties"
)
options(java.parameters = "-Xmx4g")
y <- corenlp_annotate(
  x = reuters_dt,
  destfile = NULL,
  properties_file = properties_english_fast,
  corenlp_dir = corenlp_dir,
  threads = 1L,
  progress = FALSE
  )
}
