% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{corenlp_get_properties_file}
\alias{corenlp_get_properties_file}
\title{Get path to properties file.}
\usage{
corenlp_get_properties_file(lang = c("en", "de"), fast = TRUE)
}
\arguments{
\item{lang}{the language}

\item{fast}{fast property file?}
}
\description{
The package includes minimal, adapted versions of the properties files for
English and German for testing purposes.
}
