% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{corenlp_parse_json}
\alias{corenlp_parse_json}
\title{Parse CoreNLP json string output.}
\usage{
corenlp_parse_json(
  input,
  cols_to_keep = c("sentence", "index", "word", "pos", "ner"),
  output = NULL,
  logfile = NULL,
  progress = TRUE
)
}
\arguments{
\item{input}{character vector, the JSON string(s) to be parsed}

\item{cols_to_keep}{columns to keep}

\item{output}{a destfile}

\item{logfile}{a character string naming the file to an error log to; if
provided, json strings will be written to this file if parsing the json
string string fails}

\item{progress}{logical}
}
\description{
Parse the json output from CoreNLP, either text files, or
NDJSON written to one or multiple files.
}
\details{
The JSON results of applying the Stanford CoreNLP annotators can be
written to a streaming JSON file (ndjson format). \code{corenlp_parse_json}
will parse a json string to a data.frame. If output is specified, the
output will be appended to the file provided. If \code{output} is \code{NULL}, a
data.frame is returned. Strings that cannot be parsed are written to the
logfile, if it is defined. If \code{filename} is present, the function will
process one or more files with the output of Stanford CoreNLP in a NDJSON
format. If the argument \code{output} has been defined during initialization, results are
written/appended to that file. Otherwise, a \code{data.frame} is returned.
}
