% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print.R
\name{corenlp_parse_pretty_print}
\alias{corenlp_parse_pretty_print}
\title{Parse CoreNLP pretty print output.}
\usage{
corenlp_parse_pretty_print(x, mc = 1)
}
\arguments{
\item{x}{A character vector or filename}

\item{mc}{Number of cores to use.}
}
\description{
The pretty print output of Stanford CoreNLP may be an alternative to the 
NDJSON output. The \code{corenlp_parse_pretty_print} is a residual, kept 
in the package for future use.
}
