% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purge.R
\docType{data}
\name{purge}
\alias{purge}
\alias{corenlp_postprocessing_replacements}
\alias{corenlp_preprocessing_replacements}
\title{Purge character vector from villain characters.}
\format{An object of class \code{list} of length 14.}
\usage{
purge(x, replacements, progress = TRUE)

corenlp_postprocessing_replacements

corenlp_preprocessing_replacements
}
\arguments{
\item{x}{character vector to purge}

\item{replacements}{...}

\item{progress}{logical}
}
\description{
Both preprocessing and postprocessing required!
}
\examples{
library(data.table)
options(java.parameters = "-Xmx4g")
txt <- "Karadžiš!"
corenlp_dir <- system.file(package = "cleanNLP", "extdata", "stanford-corenlp-full-2016-10-31")
corenlp_dir <- "/opt/stanford-corenlp/stanford-corenlp-full-2017-06-09"
properties_german_fast <- corenlp_get_properties_file(lang = "de")
y <- corenlp_annotate(
  x = data.table(id = 1L:length(txt), text = txt),
  properties_file = properties_german_fast, corenlp_dir = corenlp_dir,
  )
}
\keyword{datasets}
