% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purge.R
\docType{data}
\name{purge}
\alias{purge}
\alias{corenlp_postprocessing_replacements}
\alias{corenlp_preprocessing_replacements}
\title{Purge character vector from villain characters.}
\format{
An object of class \code{list} of length 14.

An object of class \code{list} of length 14.
}
\usage{
purge(x, replacements, progress = TRUE)

corenlp_postprocessing_replacements

corenlp_preprocessing_replacements
}
\arguments{
\item{x}{A \code{character} vector to purge}

\item{replacements}{A \code{list} of length-2 \code{character} vectors with a regular
expression and the corresponding replacement.}

\item{progress}{A \code{logical} value.}
}
\description{
Input and output from CoreNLP may require preprocessing and postprocessing.
\code{purge()} will also remove XML elements.
}
\examples{
library(data.table)
options(java.parameters = "-Xmx4g")
txt <- "Karadžiš!"
properties_german_fast <- corenlp_get_properties_file(lang = "de")
y <- corenlp_annotate(
  data.table(doc_id = 1L:length(txt), text = txt),
  pipe = properties_german_fast,
  corenlp_dir = corenlp_get_jar_dir(),
)
}
\keyword{datasets}
