% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment}
\alias{segment}
\title{Split table into directories with text segments.}
\usage{
segment(x, dir, chunksize = 10L, progress = interactive())
}
\arguments{
\item{x}{A \code{data.table} with columns 'doc_id' (\code{integer} values) and 'text'.
Further columns are ignored.}

\item{dir}{Superdirectory for directories with segments that will be
processed sequentially.}

\item{chunksize}{An \code{integer} value, the number of strings that will reside
in the chunk directories.}

\item{progress}{A \code{logical} value, whether to show progress bar.}
}
\value{
The function returns a \code{character} vector with the directories that
contain files with text segments.
}
\description{
Multithreaded processing using the StanfordCoreNLP class requires splitting
up input data into pieces of text ("chunks") available as files that are
processed in parallel. The \code{segment()} function performs this split operation,
i.e. it creates directories with chunks within a superdirectory.
}
\examples{
library(data.table)
reuters_txt <- readLines(system.file(package = "bignlp", "extdata", "txt", "reuters.txt"))
dt <- data.table(doc_id = 1L:length(reuters_txt), text = reuters_txt)
segdir <- tempdir()
dirs <- segment(x = dt, dir = segdir, chunksize = 10L)
}
